// src/domain/action/action-type.ts
export type ActionType =
  | 'CREATE_TRANSACTION'
  | 'UPDATE_TRANSACTION'
  | 'DELETE_TRANSACTION'
  | 'CREATE_TASK'
  | 'UPDATE_TASK'
  | 'COMPLETE_TASK'
  | 'DELETE_TASK'
  | 'CREATE_EVENT'
  | 'UPDATE_EVENT'
  | 'COMPLETE_EVENT'
  | 'DELETE_EVENT'
  | 'CREATE_NOTE'
  | 'UPDATE_NOTE'
  | 'DELETE_NOTE'
  | 'CREATE_REMINDER'
  | 'UPDATE_REMINDER'
  | 'CANCEL_REMINDER';